Abstract registration is not mandatory
"We strongly encourage pre-registering your paper by submitting its metadata (title, abstract and authors) by the corresponding deadline, which is about a month ahead of the actual paper deadline. Pre-registration is not mandatory but helps us plan the logistics. Presenters of accepted papers who pre-registered by the deadline will be rewarded with a small surprise present when they turn up in Stockholm. It is acceptable to change the metadata (title, abstract or authors) when submitting the actual paper, but then you won't get the present. For both pre-registration and submission, please visit https://eurosp19.hotcrp.com."