Added IEEE SaTML
SaTML is a conference focused on the intersection of security/privacy and machine learning. It expands on the theoretical and practical understandings of vulnerabilities inherent to machine learning (ML), explores the robustness of learning algorithms, and aids in developing a unified, coherent scientific community which aims to establish trustworthy machine learning.